[#805] Incremental standalone builds
The standalone builder keeps track of the file contents, the generated
files, emitted issues and classpath similar to what is stored in the
Eclipse IDE via the core.resources abstraction, the builder state and
the markers.

Subsequent builds will check for changed input files, missing expected
outfiles or changed dependencies and perform an incremental build
instead of a full build.

The Java-stub generation is performed incrementally, too.

Signed-off-by: Sebastian Zarnekow <sebastian.zarnekow@gmail.com>