Replace polymorphic dispatcher with EcoreSwitch
That's just a proposal (it introduces also a
org.eclipse.xtext.ecore.EcoreQualifiedNameProvider.defaultName(ENamedElement)
as in https://github.com/eclipse/xtext-extras/pull/831, and this will
probably be a conflict that should be resolved in case).

If I understand correctly, EMF switches are mostly constant time because
they perform a switch with integers.

Could it be that using the EcoreSwitch might be faster than a
polymorphic dispatch?

Signed-off-by: Lorenzo Bettini <lorenzo.bettini@gmail.com>